refactor: consolidate error classes and split server.ts
- Add shared error classes to @nikcli-ai/util (RejectedPermissionError,
  RejectedQuestionError, RejectedDBEditError, RejectedToolCallError,
  CorrectedToolCallError, DeniedToolCallError)
- Update all files to use shared error classes instead of local duplicates
- Split server.ts inline routes into separate misc.ts route file
- Remove placeholder/debug scripts from nikcli/package.json
- Add barrel export index.ts to @nikcli-ai/util
