+++ packages/nikcli/src/session/processor.ts
@@ -15,6 +15,7 @@ import { Config } from "@/config/config"
 import { SessionCompaction } from "./compaction"
 import { PermissionNext } from "@/permission/next"
 import { Question } from "@/question"
+import { RejectedPermissionError, RejectedToolCallError, RejectedQuestionError } from "@nikcli-ai/util"
 
 export namespace SessionProcessor {
   const DOOM_LOOP_THRESHOLD = 3
@@ -211,8 +212,9 @@ export namespace SessionProcessor {
                     })
 
                     if (
-                      value.error instanceof PermissionNext.RejectedError ||
-                      value.error instanceof Question.RejectedError
+                      value.error instanceof RejectedPermissionError ||
+                      value.error instanceof RejectedToolCallError ||
+                      value.error instanceof RejectedQuestionError
                     ) {
                       blocked = shouldBreak
                     }
