+++ packages/util/src/error.ts
@@ -52,3 +52,55 @@ export abstract class NamedError extends Error {
     }),
   )
 }
+
+export class RejectedPermissionError extends Error {
+  constructor(
+    public readonly sessionID: string,
+    public readonly permissionID: string,
+    public readonly toolCallID?: string,
+    public readonly metadata?: Record<string, any>,
+    public readonly reason?: string,
+  ) {
+    super(
+      reason !== undefined
+        ? reason
+        : `The user rejected permission to use this specific tool call. You may try again with different parameters.`,
+    )
+  }
+}
+
+export class RejectedQuestionError extends Error {
+  constructor() {
+    super("The user dismissed this question")
+  }
+}
+
+export class RejectedDBEditError extends Error {
+  constructor(message?: string) {
+    super(
+      message
+        ? `The user rejected database changes with feedback: ${message}`
+        : `The user rejected the database changes.`,
+    )
+  }
+}
+
+export class RejectedToolCallError extends Error {
+  constructor() {
+    super(`The user rejected permission to use this specific tool call.`)
+  }
+}
+
+export class CorrectedToolCallError extends Error {
+  constructor(message: string) {
+    super(`The user rejected permission to use this specific tool call with the following feedback: ${message}`)
+  }
+}
+
+export class DeniedToolCallError extends Error {
+  constructor(public readonly ruleset: unknown[]) {
+    super(
+      `The user has specified a rule which prevents you from using this specific tool call. Here are some of the relevant rules ${JSON.stringify(ruleset)}`,
+    )
+  }
+}
